feat: 🎸 update UI getting-started, homepage, and etc
