+++ apps/rangkaui.vercel/src/components/Header/Header.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { Flex, Text, TextInput } from 'rangkaui-next-dev';
 import { useTheme } from '@/contexts/theme';
 import Image from 'next/image';
+import Link from 'next/link';
 
 const Header = () => {
   const { backgroundColor, color } = useTheme();
@@ -12,18 +13,20 @@ const Header = () => {
       items="center"
       justify="space-between"
     >
-      <Flex className="cursor-pointer" gap={6} items="center">
-        <Image
-          className="w-[30px] h-[30px]"
-          src={'/favicon/android-chrome-192x192.png'}
-          alt="logo"
-          width={192}
-          height={192}
-        />
-        <Text fz={22} fw={700}>
-          Rangkaui
-        </Text>
-      </Flex>
+      <Link href={'/'}>
+        <Flex className="cursor-pointer" gap={6} items="center">
+          <Image
+            className="w-[30px] h-[30px]"
+            src={'/favicon/android-chrome-192x192.png'}
+            alt="logo"
+            width={192}
+            height={192}
+          />
+          <Text fz={20} fw={700}>
+            Rangkaui
+          </Text>
+        </Flex>
+      </Link>
       <Flex items="center">
         <TextInput />
       </Flex>
