+++ apps/rangkaui.vercel/src/pages/_app.tsx
@@ -1,6 +1,5 @@
 import type { AppProps } from 'next/app';
 import React from 'react';
-import MdxProvider from '@/components/Layout/Layout';
 
 import '@/styles/globals.css';
 import 'rangkaui-next/dist/styles.css';
@@ -9,9 +8,7 @@ import { ThemeProvider } from '@/contexts/theme';
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <ThemeProvider>
-      <MdxProvider>
-        <Component {...pageProps} />
-      </MdxProvider>
+      <Component {...pageProps} />
     </ThemeProvider>
   );
 }
